requirements.txt
opencv-python==4.7.0.72
mediapipe==0.10.3
tensorflow==2.11.0
scikit-learn==0.24.1
joblib==1.2.0
pandas==1.3.5
Pillow==9.5.0
numpy==1.23.5